requirements.txt
click>=8.1.0
pyyaml>=6.0
rich>=13.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
tqdm>=4.66.0
colorama>=0.4.6
flask>=3.0.0
